move serial settings from runtime to shared
